+++ chat-widget/src/components/chatbuttonstateful/ChatButtonStateful.tsx
@@ -67,7 +67,6 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
     const outOfOfficeStyleProps: IChatButtonStyleProps = Object.assign({}, defaultOutOfOfficeChatButtonStyleProps, outOfOfficeButtonProps?.styleProps);
     
     const controlProps: IChatButtonControlProps = {
-        ...buttonProps?.controlProps,
         id: "oc-lcw-chat-button",
         dir: state.domainStates.globalDir,
         titleText: "Let's Chat!",
@@ -75,8 +74,9 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
         hideNotificationBubble: buttonProps?.controlProps?.hideNotificationBubble === true || state.appStates.isMinimized === false,
         unreadMessageCount: state.appStates.unreadMessageCount ? (state.appStates.unreadMessageCount > Constants.maximumUnreadMessageCount ? props.buttonProps?.controlProps?.largeUnreadMessageString : state.appStates.unreadMessageCount.toString()) : "0",
         unreadMessageString: props.buttonProps?.controlProps?.unreadMessageString,
-        // Regular chat button onClick - this will always take precedence
-        onClick: () => ref.current()
+        // Regular chat button onClick
+        onClick: () => ref.current(),
+        ...buttonProps?.controlProps
     };
 
     const outOfOfficeControlProps: IChatButtonControlProps = {
@@ -86,14 +86,14 @@ export const ChatButtonStateful = (props: IChatButtonStatefulParams) => {
         titleText: outOfOfficeButtonProps?.controlProps?.titleText || "We're Offline",
         subtitleText: outOfOfficeButtonProps?.controlProps?.subtitleText || "No agents available",
         unreadMessageString: props.buttonProps?.controlProps?.unreadMessageString,
-        ...outOfOfficeButtonProps?.controlProps,
-        // Out-of-office specific onClick - this will ALWAYS take precedence
+        // Out-of-office specific onClick
         onClick: () => {
             if (state.appStates.isMinimized) {
                 dispatch({ type: LiveChatWidgetActionType.SET_MINIMIZED, payload: false });
             }
             dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.OutOfOffice });
-        }
+        },
+        ...outOfOfficeButtonProps?.controlProps
     };
 
     useEffect(() => {